+++ src/app/layout.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from 'next';
 import './globals.css';
 import '@/app/fonts/pretendard/font.css';
 import DeadZone from '@/components/dead-zone/DeadZone';
-import ToastContainer from '@/components/ToastContainer';
+import ToastContainer from '@/components/toast-container/ToastContainer';
 
 export const metadata: Metadata = {
   title: 'Create Next App',
